20
Ok so last push. I changed two images the 15 image and the background on the third page. Now I fixed my menu button but I didn't find a working code for the mobile responsive. And I need to find a way to connect the javascript page
and the css page. So basically I need help with some things because a lot of the code didn't work.
Still need: mobile responsive
                  fix space on my paragraphs
                  link js to css